+++ com/android/server/backup/FileMetadata.java
@@ -16,7 +16,7 @@
 
 package com.android.server.backup;
 
-import static com.android.server.backup.BackupManagerService.TAG;
+import static com.android.server.backup.RefactoredBackupManagerService.TAG;
 
 import android.app.backup.BackupAgent;
 import android.util.Slog;
@@ -36,7 +36,7 @@ public class FileMetadata {
     public long mode;                      // e.g. 0666 (actually int)
     public long mtime;                     // last mod time, UTC time_t (actually int)
     public long size;                      // bytes of content
-    public long version;                   // App version.
+    public int version;                    // App version.
     public boolean hasApk;                 // Whether backup file contains apk.
 
     @Override
