+++ com/android/server/usb/descriptors/UsbMSMidiInputJack.java
@@ -25,7 +25,7 @@ import com.android.server.usb.descriptors.report.ReportCanvas;
 public final class UsbMSMidiInputJack extends UsbACInterface {
     private static final String TAG = "UsbMSMidiInputJack";
 
-    UsbMSMidiInputJack(int length, byte type, byte subtype, int subclass) {
+    UsbMSMidiInputJack(int length, byte type, byte subtype, byte subclass) {
         super(length, type, subtype, subclass);
     }
